Codec: move to 16-bits in/out

+++ src/parameters.hh
@@ -3,8 +3,8 @@
 #pragma once
 
 constexpr struct Frame {
-  int32_t l = 0;
-  int32_t r = 0;
+  int16_t l = 0;
+  int16_t r = 0;
 } zero;
 
 constexpr int kBlockSize = 16;
